+++ Xen Tuner/xen tuner.qml
@@ -32,7 +32,7 @@ import Qt.labs.settings 1.0
 import FileIO 3.0
 
 MuseScore {
-    version: "0.4.0"
+    version: "0.4.1"
     pluginType: "dock"
     dockArea: "left"
     description: "Starts the XenTuner plugin.\n\n" +
@@ -283,27 +283,31 @@ MuseScore {
     }
 
     onScoreStateChanged: {
-        if (state.selectionChanged && curScore) {
-            var elems = curScore.selection.elements;
-            var el = elems[0];
-            var name = el ? el.name : null;
-            if (elems.length == 1 &&
-                (name == "SystemText" || name == "StaffText" ||
-                name == "TBox" || name == "Text" || name == "Fingering" ||
-                name == "Tempo" || name == "Expression")) {
-                // allow the user to use up/down arrow keys to navigate text
-                Fns.setUpDownFallthrough(false);
-                upShortcut.enabled = false;
-                downShortcut.enabled = false;
-                enharmonicShortcut.enabled = false;
-                enharmonicReversedShortcut.enabled = false;
-            } else {
-                // If no notes are selected, allow up/down arrow keys to move elements.
-                Fns.setUpDownFallthrough(true);
-                upShortcut.enabled = true;
-                downShortcut.enabled = true;
-                enharmonicShortcut.enabled = true;
-                enharmonicReversedShortcut.enabled = true;
+        if (curScore) {
+            if (state.selectionChanged) {
+                // Keep the global _curScore reference up to date.
+                Fns._curScore = curScore;
+                var elems = curScore.selection.elements;
+                var el = elems[0];
+                var name = el ? el.name : null;
+                if (elems.length == 1 &&
+                    (name == "SystemText" || name == "StaffText" ||
+                    name == "TBox" || name == "Text" || name == "Fingering" ||
+                    name == "Tempo" || name == "Expression")) {
+                    // allow the user to use up/down arrow keys to navigate text
+                    Fns.setUpDownFallthrough(false);
+                    upShortcut.enabled = false;
+                    downShortcut.enabled = false;
+                    enharmonicShortcut.enabled = false;
+                    enharmonicReversedShortcut.enabled = false;
+                } else {
+                    // If no notes are selected, allow up/down arrow keys to move elements.
+                    Fns.setUpDownFallthrough(true);
+                    upShortcut.enabled = true;
+                    downShortcut.enabled = true;
+                    enharmonicShortcut.enabled = true;
+                    enharmonicReversedShortcut.enabled = true;
+                }
             }
         }
     }